requirements.txt
blinker==1.9.0
click==8.3.3
dnspython==2.8.0
Flask==3.1.3
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
mysql-connector-python==9.7.0
psycopg2-binary==2.9.12
pymongo==4.17.0
Werkzeug==3.1.8
